[IG] Terminé los tests de VehiculoController
Me falta ver cómo hacer los otros 2

+++ src/main/java/usach/tingeso/controllers/BoletaController.java
@@ -2,8 +2,10 @@ package usach.tingeso.controllers;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import usach.tingeso.entities.BoletaEntity;
 import usach.tingeso.entities.ReparacionEntity;
 import usach.tingeso.services.BoletaService;
@@ -21,23 +23,41 @@ public class BoletaController {
 
     @GetMapping("/")
     public ResponseEntity<List<BoletaEntity>> getBoletas(){
-        return ResponseEntity.ok(boletaService.getBoletas());
+        List<BoletaEntity> boletas = boletaService.getBoletas();
+        if(boletas == null){
+            throw new ResponseStatusException(org.springframework.http.HttpStatus.NOT_FOUND, "Boleta not found");
+        }
+        return ResponseEntity.ok(boletas);
     }
     @GetMapping("/{reparacion}")
     public ResponseEntity<BoletaEntity> getBoletaByReparacion(@PathVariable ReparacionEntity reparacion){
-        return ResponseEntity.ok(boletaService.getBoletaByReparacion(reparacion));
+        BoletaEntity boleta = boletaService.getBoletaByReparacion(reparacion);
+        if(boleta == null){
+            throw new ResponseStatusException(org.springframework.http.HttpStatus.NOT_FOUND, "Boleta not found");
+        }
+        return ResponseEntity.ok(boleta);
     }
     @PostMapping("/")
     public ResponseEntity<BoletaEntity> saveBoleta(@RequestBody BoletaEntity boleta){
 
+        if(boleta==null){
+            throw new ResponseStatusException(org.springframework.http.HttpStatus.BAD_REQUEST, "Boleta already exists");
+        }
         return ResponseEntity.ok(boletaService.saveBoleta(boleta));
     }
     @PutMapping("/")
     public ResponseEntity<BoletaEntity> updateBoleta(@RequestBody BoletaEntity boleta){
+        if(boleta==null){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "NOT FOUND");
+        }
         return ResponseEntity.ok(boletaService.saveBoleta(boleta));
     }
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteBoleta(@PathVariable Long id){
+        BoletaEntity boleta = boletaService.getBoletaById(id);
+        if(boleta == null){
+            throw new ResponseStatusException(org.springframework.http.HttpStatus.NOT_FOUND, "Boleta not found");
+        }
         return ResponseEntity.ok(boletaService.deleteBoleta(id));
     }
 
@@ -65,6 +85,9 @@ public class BoletaController {
 //    }
     @PutMapping("/init/")
     public ResponseEntity<?> inicializarBoleta(@RequestBody BoletaEntity boleta){
+        if(boleta==null){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Boleta already exists");
+        }
         boletaService.guardarRecargoPorKM(boleta);
         boletaService.guardarRecargoPorAntiguedad(boleta);
         boletaService.guardarDescuentoPorDia(boleta);
